+++ server/backends/systemInfo/systemInfo.php
@@ -1,0 +1,23 @@
+<?php
+
+    /**
+     * backends systemInfo namespace
+     */
+
+    namespace backends\systemInfo {
+
+        use backends\backend;
+
+        /**
+         * base systemInfo class
+         */
+
+        abstract class systemInfo extends backend {
+
+            /**
+             * @return array
+             */
+
+            abstract public function systemInfo();
+        }
+    }