+++ changelog.md
@@ -8,6 +8,7 @@
 ```
 
 ```diff
++ simple system info dashboard
 ! backend tt type "mongo" renamed to "internal" (need modify server/config/config.json)
 + camTree settings for webUI, camTree = false - off, "houses" - common for all houses, "perHouse" - per house
 + devices (cameras and domophones) tree